phase2_resume_extraction/requirements.txt
pypdf
python-docx
pydantic
